fixed creative mode checks and indentation
Simplified one of the creative mode break checks to remove unnecessary duplicate code. Also added a creative mode check so that using tools via interaction would always work on obsidian brick in creative mode, since the hurt method only covers direct attacks. It wouldn't make sense to be able to destroy it directly but not with a tool.